Made factory pattern for TextBoxes and adjusted all positions of Non-…

+++ src/main/java/GameDoneView.java
@@ -107,7 +107,7 @@ public class GameDoneView extends BasicGameState {
     private void startMusic(){
         try{
             Sound anthem = new Sound("data/music/Stad_i_ljus.wav");
-            anthem.play(1, 1);
+            anthem.play((float)1, 1);
             musicStarted = true;
         }catch(SlickException e){
             e.printStackTrace();
